requirements.txt
click==8.0.3
colorama==0.4.4
dominate==2.6.0
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
freeopcua==0.90.6
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.6.2
MarkupSafe==2.0.1
modbus-tk==1.1.2
opcua==0.98.12
opcua-client==0.8.0
opcua-widgets==0.5.10
OpenOPC-Python3x==1.3.1
PyQt5==5.15.2
PyQt5-sip==12.8.1
Pyro4==4.80
pyserial==3.5
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
serpent==1.30.2
six==1.15.0
visitor==0.1.3
Werkzeug==2.0.3
